Move function to generate templates in random order into Bams and add test.
